Reworking mocks classes
